warning fix - some missing cast is written, a useless variable is removed

+++ src/editor.cpp
@@ -34,7 +34,7 @@
 Editor* Editor::m_editor = 0;
 
 // ----------------------------------------------------------------------------
-bool Editor::buttonClicked(int ID)
+bool Editor::buttonClicked(s32 ID)
 {
     RoadPanel* rp;
     switch (ID)
@@ -183,7 +183,7 @@ bool Editor::buttonClicked(int ID)
     }
 
     if (ID >= m_tex_sel->FIRST_BTN_ID &&
-        ID < m_tex_sel->FIRST_TEX_BTN_ID + m_tex_sel->getBtnNum())
+        ID < m_tex_sel->FIRST_TEX_BTN_ID + (s32)m_tex_sel->getBtnNum())
     {
         m_tex_sel->btnClicked(ID);
         return true;
@@ -617,7 +617,7 @@ bool Editor::run()
 
         m_viewport->draw();
         m_rcs->render();
-        
+
 		m_gui_env->drawAll();
 
         if (m_viewport->getState() != Viewport::FREECAM && !m_rcs->isVisible())
@@ -697,7 +697,7 @@ bool Editor::OnEvent(const SEvent& event)
         // check if mouse is outside of the viewport's domain
         if (event.MouseInput.Y < 50 || event.MouseInput.X <= 15 ||
             event.MouseInput.X >(s32) m_screen_size.Width - 250 ||
-             event.MouseInput.Y >= m_screen_size.Height - 15 ||
+             event.MouseInput.Y >= (s32)m_screen_size.Height - 15 ||
             (event.MouseInput.X >(s32)m_screen_size.Width - 500 && m_tex_sel->isActive()))
         {
             u32 s = m_viewport->getState();
@@ -859,11 +859,11 @@ void Editor::keepMouseIn(s32 sx, s32 sy)
     s32 y = sy;
     if (sy < 50)
         y = m_screen_size.Height - sy;
-    else if (sy >= m_screen_size.Height - 15)
+    else if (sy >= (s32)m_screen_size.Height - 15)
         y = 55;
     if (sx <= 15)
-        x = m_screen_size.Width - 250;
-    else if (sx >= m_screen_size.Width - 250)
+        x = (s32)m_screen_size.Width - 250;
+    else if (sx >= (s32)m_screen_size.Width - 250)
         x = 20;
     m_device->getCursorControl()->setPosition(x, y);
     m_mouse.prev_x = x;
